Extract the worker-side slice pipeline into amrexplorer_pipeline
Move the slice orchestration out of MainWindow's anonymous namespace into the
Qt-free pipeline library: executeSlice, appendContours, appendVectorGlyphs,
refreshCachedSlice, recomputeContourPolylines, executeFrameLoad, the
geometry/level helpers, and the SliceDisplayResult/InitialSliceResult/
FrameSliceSpec/DisplayMode types, alongside the range-resolution cluster
already in SliceRangeResolver. Qt-side names keep resolving through
using-declarations.

Qt dependencies were cut at three seams: executeFrameLoad takes the cache
budget as a parameter instead of reading the environment, its cache-pressure
errors are plain strings the GUI wraps in translated text, and the GUI keeps a
QString wrapper over cacheBudgetDescription.

Add unit tests for the range resolver: masking/padding, metadata-range
union/exact/missing-statistics, the Visible downgrade, degenerate and throwing
ranges, the log-to-linear fallback, and that the User/Visible paths never
touch the dataset.

MainWindow.cpp shrinks 5,835 to 5,197 lines. No behavior change.
