+++ src/qt/MainWindow.cpp
@@ -17,6 +17,7 @@
 #include <amrexplorer/io/FabCatalog.hpp>
 #include <amrexplorer/io/StandaloneMetadataReader.hpp>
 #include <amrexplorer/core/Statistics.hpp>
+#include <amrexplorer/pipeline/SliceRangeResolver.hpp>
 #include <amrexplorer/query/LineQuery.hpp>
 #include <amrexplorer/query/SliceQuery.hpp>
 #include <amrexplorer/render2d/Contours.hpp>
@@ -239,530 +240,11 @@ QString exceptionMessage(const std::exception& error)
     return QString::fromUtf8(error.what());
 }
 
-std::optional<std::pair<double, double>> finiteRange(
-    const ScalarPlane& plane)
-{
-    auto minimum = std::numeric_limits<double>::infinity();
-    auto maximum = -std::numeric_limits<double>::infinity();
-    for (std::size_t pixel = 0; pixel < plane.values.size(); ++pixel) {
-        if (plane.valid[pixel] == 0) {
-            continue;
-        }
-        const auto value = static_cast<double>(plane.values[pixel]);
-        if (!std::isfinite(value)) {
-            continue;
-        }
-        minimum = std::min(minimum, value);
-        maximum = std::max(maximum, value);
-    }
-    if (!std::isfinite(minimum) || !std::isfinite(maximum)) {
-        return std::nullopt;
-    }
-    if (minimum == maximum) {
-        const auto padding = std::max(std::abs(minimum), 1.0) * 1.0e-6;
-        minimum -= padding;
-        maximum += padding;
-    }
-    return std::pair{minimum, maximum};
-}
-
-std::optional<ValueRange> selectedMetadataRange(
-    const DatasetMetadata& metadata, FieldId field, int maximumLevel,
-    CompositionPolicy composition, RangeMode rangeMode)
-{
-    if (rangeMode == RangeMode::File) {
-        return metadataValueRange(metadata, field, std::nullopt);
-    }
-    if (rangeMode != RangeMode::Level) {
-        return std::nullopt;
-    }
-    if (composition == CompositionPolicy::ExactLevel) {
-        return metadataValueRange(metadata, field, maximumLevel);
-    }
-
-    auto minimum = std::numeric_limits<double>::infinity();
-    auto maximum = -std::numeric_limits<double>::infinity();
-    for (int level = 0; level <= maximumLevel; ++level) {
-        const auto range = metadataValueRange(metadata, field, level);
-        if (!range) {
-            return std::nullopt;
-        }
-        minimum = std::min(minimum, range->minimum);
-        maximum = std::max(maximum, range->maximum);
-    }
-    if (!std::isfinite(minimum) || !std::isfinite(maximum)) {
-        return std::nullopt;
-    }
-    return ValueRange{minimum, maximum};
-}
-
-RangeMode effectiveRangeMode(
-    const DatasetMetadata& metadata, FieldId field, int maximumLevel,
-    CompositionPolicy composition, RangeMode requested)
-{
-    if (metadata.isFab && requested == RangeMode::File) {
-        return requested;
-    }
-    if ((requested == RangeMode::File || requested == RangeMode::Level)
-        && !selectedMetadataRange(
-            metadata, field, maximumLevel, composition, requested)) {
-        return RangeMode::Visible;
-    }
-    return requested;
-}
-
-std::optional<std::pair<double, double>> fabDataRange(
-    const std::shared_ptr<PlotfileDataset>& dataset, FieldId field)
-{
-    if (!dataset->metadata().isFab) {
-        return std::nullopt;
-    }
-    BlockRequest request;
-    request.dataset = dataset->id();
-    request.field = field;
-    const auto access = dataset->requestBlock(request);
-    const auto& values = access.handle->values;
-    auto minimum = std::numeric_limits<double>::infinity();
-    auto maximum = -std::numeric_limits<double>::infinity();
-    for (std::size_t index = 0; index < values.size(); ++index) {
-        const auto value = values[index];
-        if (!std::isfinite(value)) {
-            continue;
-        }
-        minimum = std::min(minimum, value);
-        maximum = std::max(maximum, value);
-    }
-    if (!std::isfinite(minimum) || !std::isfinite(maximum)) {
-        return std::nullopt;
-    }
-    return std::pair{minimum, maximum};
-}
-
-// The display range for a slice: the user's explicit range, the level/file
-// metadata range, or the finite extrema of the plane itself, padded so
-// minimum < maximum always holds. A logarithmic request whose range is not
-// strictly positive throws, so the caller can fall back to linear. Shared by
-// executeSlice and the re-render-from-cache path, which must agree exactly.
-std::pair<double, double> resolveRange(
-    const std::shared_ptr<PlotfileDataset>& dataset, FieldId field,
-    int maximumLevel, CompositionPolicy composition, RangeMode rangeMode,
-    const std::optional<std::pair<double, double>>& userRange,
-    bool logarithmic, const ScalarPlane& plane)
-{
-    auto selectedRange = userRange;
-    if (rangeMode == RangeMode::Level || rangeMode == RangeMode::File) {
-        if (rangeMode == RangeMode::File) {
-            selectedRange = fabDataRange(dataset, field);
-        }
-        if (!selectedRange) {
-            const auto statistics = selectedMetadataRange(dataset->metadata(),
-                field, maximumLevel, composition, rangeMode);
-            if (statistics) {
-                selectedRange = std::pair{
-                    statistics->minimum, statistics->maximum};
-            }
-        }
-    }
-    auto [minimum, maximum] = selectedRange
-        ? *selectedRange
-        : finiteRange(plane).value_or(logarithmic
-              ? std::pair{1.0, 10.0}
-              : std::pair{0.0, 1.0});
-    if (minimum == maximum) {
-        if (logarithmic && minimum > 0.0) {
-            minimum /= 1.0 + 1.0e-6;
-            maximum *= 1.0 + 1.0e-6;
-        } else {
-            const auto padding = std::max(std::abs(minimum), 1.0) * 1.0e-6;
-            minimum -= padding;
-            maximum += padding;
-        }
-    }
-    if (!(minimum < maximum)) {
-        throw std::runtime_error("user scalar range must have positive extent");
-    }
-    if (logarithmic && !(minimum > 0.0)) {
-        throw std::runtime_error("logarithmic scalar range must be positive");
-    }
-    return {minimum, maximum};
-}
-
-// Upper bound on slice output dimensions. One pixel per finest cell is the
-// ideal, but a plane costs on the order of 10 bytes per pixel (float value,
-// validity mask, source level, RGBA image), so uncapped native resolution on
-// huge domains could allocate gigabytes. Once the region edges are snapped
-// to cell boundaries (snapToCellBoundaries) the extent is an exact multiple
-// of the cell size, so when this cap does engage the sampling pitch only
-// ever exceeds the cell size — honest downsampling — and never produces
-// duplicated or skipped cells.
-constexpr int maxSliceOutputDimension = 4096;
-
-// Native render resolution for a slice: the count of finest-level cells the
-// visible region spans along each in-plane axis. At the 1x fixed scale this is
-// the resolution legacy Amrvis drew (one pixel per finest cell); the larger
-// fixed scales magnify it through the view zoom.
-std::array<int, 2> finestNativeOutputSize(
-    const amrvis::DatasetMetadata& metadata, const amrvis::RealBox& region,
-    int normal)
-{
-    const auto& finest = metadata.levels[static_cast<std::size_t>(
-        std::max(0, metadata.finestLevel))];
-    std::array<int, 2> axes{0, 1};
-    if (metadata.dimension == 3) {
-        std::size_t next = 0;
-        for (int axis = 0; axis < 3; ++axis) {
-            if (axis != normal) {
-                axes[next++] = axis;
-            }
-        }
-    }
-    const auto cells = [&](int axis) {
-        const auto i = static_cast<std::size_t>(axis);
-        const auto extent = region.upper[i] - region.lower[i];
-        return std::clamp(
-            static_cast<int>(std::round(extent / finest.cellSize[i])),
-            1, maxSliceOutputDimension);
-    };
-    return {cells(axes[0]), cells(axes[1])};
-}
-
-// Like resolveRange, but if a logarithmic scale is requested and the range
-// is not strictly positive (a Visible/Level/File/User minimum <= 0), it falls
-// back to a linear range and reports logarithmic=false so the caller renders
-// linearly instead of failing the whole slice. A slice with no finite values
-// uses a neutral positive range and can therefore remain logarithmic.
-struct ResolvedRange {
-    double minimum;
-    double maximum;
-    bool logarithmic;
-};
-
-ResolvedRange resolveDisplayRange(
-    const std::shared_ptr<PlotfileDataset>& dataset, FieldId field,
-    int maximumLevel, CompositionPolicy composition, RangeMode rangeMode,
-    const std::optional<std::pair<double, double>>& userRange,
-    bool logarithmic, const ScalarPlane& plane)
-{
-    if (logarithmic) {
-        try {
-            const auto [minimum, maximum] = resolveRange(dataset, field,
-                maximumLevel, composition, rangeMode, userRange, true, plane);
-            return {minimum, maximum, true};
-        } catch (const std::exception&) {
-            // Log is not viable for this range; fall back to linear below.
-        }
-    }
-    const auto [minimum, maximum] = resolveRange(dataset, field, maximumLevel,
-        composition, rangeMode, userRange, false, plane);
-    return {minimum, maximum, false};
-}
-
-SliceDisplayResult executeSlice(const std::shared_ptr<PlotfileDataset>& dataset,
-    const SliceRequest& request,
-    RangeMode rangeMode,
-    const std::optional<std::pair<double, double>>& userRange,
-    bool logarithmic, const Palette& palette, StopToken cancellation)
-{
-    SliceDisplayResult result;
-    result.request = request;
-    result.slice = SliceQuery(*dataset).execute(request, cancellation);
-    const auto range = resolveDisplayRange(dataset, request.field,
-        request.maximumLevel, request.composition, rangeMode, userRange,
-        logarithmic, result.slice.plane);
-    result.minimum = range.minimum;
-    result.maximum = range.maximum;
-    result.logarithmic = range.logarithmic;
-    result.fieldName = dataset->metadata().fields[request.field.value].name;
-    result.image = renderScalarPlane(result.slice.plane,
-        ScalarRenderSettings{
-            .minimum = range.minimum,
-            .maximum = range.maximum,
-            .logarithmic = range.logarithmic,
-            .palette = &palette
-        });
-    return result;
-}
-
-// Vector mode queries the U- and V-component planes independently and
-// derives arrow glyphs from them. Both slices share the raster request's
-// region, level, and output size so the planes line up sample for sample.
-void appendVectorGlyphs(const std::shared_ptr<PlotfileDataset>& dataset,
-    SliceRequest request, FieldId uField, FieldId vField, int count,
-    StopToken cancellation, SliceDisplayResult& result)
-{
-    request.field = uField;
-    auto uSlice = SliceQuery(*dataset).execute(request, cancellation);
-    request.field = vField;
-    auto vSlice = SliceQuery(*dataset).execute(request, cancellation);
-    result.vectors = generateVectorGlyphs(uSlice.plane, vSlice.plane, count);
-    result.slice.metrics.candidateBlocks += uSlice.metrics.candidateBlocks
-        + vSlice.metrics.candidateBlocks;
-    result.slice.metrics.blocksRead += uSlice.metrics.blocksRead
-        + vSlice.metrics.blocksRead;
-    result.slice.metrics.cacheHits += uSlice.metrics.cacheHits
-        + vSlice.metrics.cacheHits;
-    result.slice.metrics.payloadBytesRead += uSlice.metrics.payloadBytesRead
-        + vSlice.metrics.payloadBytesRead;
-}
-
-[[nodiscard]] bool isContourMode(DisplayMode mode)
-{
-    return mode == DisplayMode::RasterContours;
-}
-
-// The cache-key comparison for PlaneViewState: everything a cached slice
-// depends on. Range, log scale, palette, and contour count are deliberately
-// absent — those are recomputed from the cached planes on the cheap path.
-[[nodiscard]] bool sameSliceSpec(const SliceRequest& lhs, const SliceRequest& rhs)
-{
-    return lhs.dataset == rhs.dataset && lhs.field == rhs.field
-        && lhs.component == rhs.component
-        && lhs.normalDirection == rhs.normalDirection
-        && lhs.physicalPosition == rhs.physicalPosition
-        && lhs.visibleRegion == rhs.visibleRegion
-        && lhs.maximumLevel == rhs.maximumLevel
-        && lhs.outputSize == rhs.outputSize
-        && lhs.sampling == rhs.sampling
-        && lhs.composition == rhs.composition;
-}
-
-// The two in-plane axes of a slice, mirroring SliceQuery's plane axes.
-[[nodiscard]] std::array<int, 2> slicePlaneAxes(int dimension, int normalDirection)
-{
-    if (dimension == 2) {
-        return {0, 1};
-    }
-    std::array<int, 2> axes{};
-    std::size_t next = 0;
-    for (int axis = 0; axis < 3; ++axis) {
-        if (axis != normalDirection) {
-            axes[next++] = axis;
-        }
-    }
-    return axes;
-}
-
-// The number of cells of `level` covering [lower, upper] on `axis`, clipped
-// to the level's index domain: the data resolution of a slice request.
-[[nodiscard]] int coveredCells(const DatasetMetadata& metadata, int level,
-    int axis, double lower, double upper)
-{
-    const auto& levelMetadata = metadata.levels[static_cast<std::size_t>(level)];
-    const auto index = static_cast<std::size_t>(axis);
-    const auto bounds = sampleBounds(
-        levelMetadata, levelMetadata.domain, metadata.dimension);
-    const auto lo = std::max(lower, bounds.lower[index]);
-    const auto hi = std::min(upper, bounds.upper[index]);
-    if (!(lo < hi)) {
-        return 1;
-    }
-    const auto domainCells = static_cast<std::int64_t>(
-        levelMetadata.domain.upper[index]) - levelMetadata.domain.lower[index];
-    const auto first = std::clamp<std::int64_t>(
-        static_cast<std::int64_t>(sampleIndex(levelMetadata, axis, lo))
-            - levelMetadata.domain.lower[index],
-        0, domainCells);
-    const auto last = std::clamp<std::int64_t>(
-        static_cast<std::int64_t>(sampleIndex(
-            levelMetadata, axis, std::nextafter(hi, lo)))
-            - levelMetadata.domain.lower[index],
-        0, domainCells);
-    return static_cast<int>(last - first + 1);
-}
-
-// Contour overlays are extracted from a piecewise-constant slice at DATA
-// resolution: one sample per cell of the finest participating level covering
-// the visible region, capped at 1024 samples per axis (for coarse data this
-// plane is tiny — 4x4 on the test fixture). When the data is finer than the
-// display on both axes the display plane doubles as the contour plane,
-// because the cell-scale staircase is sub-pixel there. The contour plane is
-// then bilinearly refined so one fine cell spans at most a few display
-// pixels; marching squares on the refinement converges to the exact
-// iso-curve of the bilinear interpolant (no new extrema, no ringing), and a
-// single Chaikin pass finishes the polyline. This replaces the old second
-// SliceQuery with linear sampling at display resolution, which cost about as
-// much as the display query itself. Extraction runs here, off the GUI
-// thread, with the output mapped to display-plane pixel space; the GUI
-// thread only converts the polylines to painter paths in updateOverlay. The
-// planes and the refinement are cached in PlaneViewState, so range and
-// contour-count changes re-run only this cheap extraction (see
-// refreshCachedSlice).
-void appendContours(const std::shared_ptr<PlotfileDataset>& dataset,
-    const SliceRequest& request, int contourCount, double minimum, double maximum,
-    bool logarithmic, StopToken cancellation, SliceDisplayResult& result)
-{
-    const auto& metadata = dataset->metadata();
-    const auto level = std::min(request.maximumLevel, metadata.finestLevel);
-    const auto axes = slicePlaneAxes(metadata.dimension, request.normalDirection);
-    const auto xAxis = static_cast<std::size_t>(axes[0]);
-    const auto yAxis = static_cast<std::size_t>(axes[1]);
-    const auto dataWidth = coveredCells(metadata, level, axes[0],
-        request.visibleRegion.lower[xAxis], request.visibleRegion.upper[xAxis]);
-    const auto dataHeight = coveredCells(metadata, level, axes[1],
-        request.visibleRegion.lower[yAxis], request.visibleRegion.upper[yAxis]);
-    const auto displayWidth = request.outputSize[0];
-    const auto displayHeight = request.outputSize[1];
-
-    // Legacy Amrvis draws contours at each FAB's native grid resolution,
-    // producing smooth curves at any display scale. We match that by
-    // querying a linearly interpolated plane at a resolution fine enough
-    // that contour staircases are invisible — at least 512 samples on the
-    // shorter axis, capped at 1024. Two Chaikin passes finish the polylines.
-    auto contourRequest = request;
-    contourRequest.outputSize = {
-        std::min(std::max(dataWidth, 512), 1024),
-        std::min(std::max(dataHeight, 512), 1024)};
-    contourRequest.sampling = SamplingPolicy::Linear;
-    auto contour = SliceQuery(*dataset).execute(contourRequest, cancellation);
-    result.contourPlane = std::move(contour.plane);
-    result.slice.metrics.candidateBlocks += contour.metrics.candidateBlocks;
-    result.slice.metrics.blocksRead += contour.metrics.blocksRead;
-    result.slice.metrics.cacheHits += contour.metrics.cacheHits;
-    result.slice.metrics.payloadBytesRead += contour.metrics.payloadBytesRead;
-
-    // No supersampling: the linear plane is already smooth.  Store it as
-    // the fine plane too so refreshCachedSlice can reuse it.
-    result.contourFinePlane = result.contourPlane;
-    result.contourFineFactor = 1;
-    const auto values = contourValues(
-        minimum, maximum, contourCount, logarithmic);
-    result.contourPolylines = contourPolylinesForDisplay(
-        result.contourFinePlane, 1, values, displayWidth, displayHeight);
-}
-
-// Re-render-from-cache: only palette/log/range/contour-count cosmetics
-// changed (the request still matches the view's cache key), so the cached
-// planes are re-ranged, re-rendered, and re-contoured without any SliceQuery.
-// With rasterDirty false the raster is known unchanged and the image is not
-// re-rendered; SliceDisplayResult::rasterUnchanged tells showSlice to keep
-// the view's pixmap. Vector glyphs are reused from the cache: they do not
-// depend on palette/log/range. Runs on a worker; delivery uses the same
-// watcher and generation flow as a full slice request.
-SliceDisplayResult refreshCachedSlice(
-    const std::shared_ptr<PlotfileDataset>& dataset,
-    const SliceRequest& request, ScalarPlane displayPlane,
-    ScalarPlane contourPlane, ScalarPlane contourFinePlane, int contourFineFactor,
-    std::vector<VectorSegment> vectors,
-    RangeMode rangeMode,
-    const std::optional<std::pair<double, double>>& userRange,
-    bool logarithmic, const Palette& palette, DisplayMode displayMode,
-    std::uint32_t vectorUField, std::uint32_t vectorVField,
-    int contourCount, bool rasterDirty)
-{
-    SliceDisplayResult result;
-    result.request = request;
-    result.mode = displayMode;
-    result.vectorUField = vectorUField;
-    result.vectorVField = vectorVField;
-    result.contourCount = contourCount;
-    result.slice.plane = std::move(displayPlane);
-    const auto range = resolveDisplayRange(dataset, request.field,
-        request.maximumLevel, request.composition, rangeMode, userRange,
-        logarithmic, result.slice.plane);
-    result.minimum = range.minimum;
-    result.maximum = range.maximum;
-    result.logarithmic = range.logarithmic;
-    result.fieldName = dataset->metadata().fields[request.field.value].name;
-    result.rasterUnchanged = !rasterDirty;
-    if (rasterDirty) {
-        result.image = renderScalarPlane(result.slice.plane,
-            ScalarRenderSettings{
-                .minimum = range.minimum,
-                .maximum = range.maximum,
-                .logarithmic = range.logarithmic,
-                .palette = &palette
-            });
-    }
-    if (isContourMode(displayMode)) {
-        result.contourPlane = std::move(contourPlane);
-        result.contourFinePlane = std::move(contourFinePlane);
-        result.contourFineFactor = contourFineFactor;
-        const auto values = contourValues(
-            range.minimum, range.maximum, contourCount, range.logarithmic);
-        result.contourPolylines = contourPolylinesForDisplay(
-            result.contourFinePlane, contourFineFactor, values,
-            request.outputSize[0], request.outputSize[1]);
-    }
-    if (displayMode == DisplayMode::VelocityVectors) {
-        result.vectors = std::move(vectors);
-    }
-    return result;
-}
-
-// Re-extract contour polylines from an already-populated fine plane after the
-// display range is replaced downstream of appendContours/refreshCachedSlice —
-// full-domain-range reuse, the 3-D shared Visible range, or syncVisibleRanges.
-// Cheap: the fine plane is cached, so no SliceQuery runs. Returns empty when
-// the range is unusable (non-positive log range, zero extent), matching
-// updateOverlay's guard, so a bad range clears rather than throws. See the
-// contours-stale-after-visible-range-sync issue.
-std::vector<ContourPolyline> recomputeContourPolylines(
-    const ScalarPlane& finePlane, int fineFactor, double minimum,
-    double maximum, bool logarithmic, int contourCount,
-    int displayWidth, int displayHeight)
-{
-    if (finePlane.width <= 0 || finePlane.height <= 0 || contourCount < 1
-        || !(minimum < maximum) || (logarithmic && !(minimum > 0.0))) {
-        return {};
-    }
-    try {
-        const auto values = contourValues(
-            minimum, maximum, contourCount, logarithmic);
-        return contourPolylinesForDisplay(
-            finePlane, fineFactor, values, displayWidth, displayHeight);
-    } catch (const std::exception&) {
-        return {};
-    }
-}
-
-// SliceDisplayResult overload: regenerate the polylines to match the result's
-// current (possibly replaced) minimum/maximum. No-op outside contour modes.
-void recomputeContourPolylines(SliceDisplayResult& result)
-{
-    if (!isContourMode(result.mode)) {
-        return;
-    }
-    result.contourPolylines = recomputeContourPolylines(
-        result.contourFinePlane, result.contourFineFactor, result.minimum,
-        result.maximum, result.logarithmic, result.contourCount,
-        result.request.outputSize[0], result.request.outputSize[1]);
-}
-
-// Combo data sentinel for "Update to Level N" entries, which composite
-// levels 0..N with FinestAvailable. The selected level N is data - 1000.
-constexpr int kUpdateToLevelOffset = 1000;
-
-struct LevelSelection {
-    CompositionPolicy composition = CompositionPolicy::FinestAvailable;
-    int maximumLevel = 0;
-};
-
-LevelSelection decodeLevelData(int data, int finestLevel)
-{
-    if (data >= kUpdateToLevelOffset) {
-        return {CompositionPolicy::FinestAvailable, data - kUpdateToLevelOffset};
-    }
-    if (data < 0) {
-        return {CompositionPolicy::FinestAvailable, finestLevel};
-    }
-    return {CompositionPolicy::ExactLevel, data};
-}
-
+// QString face of the pipeline's formatter, for the GUI-side messages; hides
+// amrvis::cacheBudgetDescription for unqualified calls in this namespace.
 QString cacheBudgetDescription(std::uint64_t bytes)
 {
-    constexpr std::uint64_t kibibyte = 1024;
-    constexpr std::uint64_t mebibyte = 1024 * kibibyte;
-    constexpr std::uint64_t gibibyte = 1024 * mebibyte;
-    if (bytes % gibibyte == 0) {
-        return QObject::tr("%1 GiB").arg(bytes / gibibyte);
-    }
-    if (bytes % mebibyte == 0) {
-        return QObject::tr("%1 MiB").arg(bytes / mebibyte);
-    }
-    if (bytes % kibibyte == 0) {
-        return QObject::tr("%1 KiB").arg(bytes / kibibyte);
-    }
-    return QObject::tr("%1 bytes").arg(bytes);
+    return QString::fromStdString(amrvis::cacheBudgetDescription(bytes));
 }
 
 QString cacheFallbackMessage(
@@ -816,228 +298,6 @@ void populateLevelCombo(QComboBox* combo, int finestLevel)
     }
 }
 
-// How a slice-plane index maps to a physical coordinate. Indices use the
-// level's integer box space (domain.lower .. domain.upper), which can be
-// negative. Cell-centered data places the value at the center of each cell
-// (index i → prob_lo + (i - domain.lower + 0.5)*dx); nodal data places it
-// at the node (index i → prob_lo + (i - domain.lower)*dx).
-int sliceIndexForPosition(const DatasetMetadata& md, int level, int axis,
-    double position)
-{
-    const auto& levelMd = md.levels[static_cast<std::size_t>(level)];
-    return sampleIndex(levelMd, axis, position);
-}
-
-double positionForSliceIndex(const DatasetMetadata& md, int level, int axis,
-    int index)
-{
-    const auto& levelMd = md.levels[static_cast<std::size_t>(level)];
-    return samplePosition(levelMd, axis, index);
-}
-
-// Opens one plotfile on a worker thread and renders the slice(s) described
-// by spec — one per ortho view for 3-D, the single y-normal view for 2-D.
-// Shared by the initial open path (default spec) and the sequence path
-// (spec preserving the user's UI state across frames).
-InitialSliceResult executeFrameLoad(const std::filesystem::path& path,
-    DatasetId datasetId, const FrameSliceSpec& spec, StopToken cancellation,
-    std::optional<PlotfileMetadataResult> preparedMetadata = std::nullopt,
-    std::filesystem::path dataRoot = {})
-{
-    InitialSliceResult result;
-    const auto cacheBudget = initialCacheBudget();
-    if (preparedMetadata) {
-        result.fileVersion = preparedMetadata->fileVersion;
-        result.dataset = std::make_shared<PlotfileDataset>(
-            std::move(dataRoot), datasetId, cacheBudget,
-            std::move(*preparedMetadata));
-    } else {
-        result.dataset = std::make_shared<PlotfileDataset>(
-            path, datasetId, cacheBudget);
-    }
-    const auto& metadata = result.dataset->metadata();
-    if (metadata.fields.empty()) {
-        throw std::runtime_error("dataset has no scalar fields to display");
-    }
-    if (result.fileVersion.empty()) {
-        std::ifstream header(path / "Header");
-        std::getline(header, result.fileVersion);
-        while (!result.fileVersion.empty() && result.fileVersion.back() == '\r') {
-            result.fileVersion.pop_back();
-        }
-    }
-
-    const auto fieldCount = static_cast<std::uint32_t>(metadata.fields.size());
-    const auto field = std::min(spec.field, fieldCount - 1);
-    // An out-of-range exact level falls back to finest-available, matching
-    // the level combo's behavior when a frame has fewer levels.
-    // Combo data encoding: -1=finest, N=level N only, 1000+N=update to N.
-    const auto levelSelection = spec.levelSelection >= -1
-        && spec.levelSelection <= metadata.finestLevel
-            ? spec.levelSelection
-            : (spec.levelSelection >= kUpdateToLevelOffset
-                && spec.levelSelection - kUpdateToLevelOffset
-                    <= metadata.finestLevel)
-            ? spec.levelSelection
-            : -1;
-    const auto selectedLevel = decodeLevelData(
-        levelSelection, metadata.finestLevel);
-    auto attemptMaximumLevel = selectedLevel.maximumLevel;
-    const auto rangeMode = effectiveRangeMode(metadata, FieldId{field},
-        attemptMaximumLevel, selectedLevel.composition, spec.rangeMode);
-    std::array<double, 3> positions{0.0, 0.0, 0.0};
-    const auto dataBounds = datasetSampleBounds(metadata);
-    for (std::size_t axis = 0; axis < 3; ++axis) {
-        const auto lower = dataBounds.lower[axis];
-        const auto upper = dataBounds.upper[axis];
-        positions[axis] = spec.defaultPositions
-            ? lower + 0.5 * (upper - lower)
-            : std::clamp(spec.slicePositions[axis], lower,
-                std::nextafter(upper, lower));
-    }
-
-    // 3-D datasets display all three orthogonal planes at once; the slices
-    // share the dataset cache, so the sequential queries are bounded. 2-D
-    // keeps its single y-normal display plane.
-    const std::vector<int> normals = metadata.dimension == 3
-        ? std::vector<int>{0, 1, 2} : std::vector<int>{1};
-    for (;;) {
-        try {
-            result.displays.clear();
-            result.displays.reserve(normals.size());
-            for (std::size_t entry = 0; entry < normals.size(); ++entry) {
-                const auto normal = normals[entry];
-                SliceRequest request;
-                request.dataset = datasetId;
-                request.field = FieldId{field};
-                request.normalDirection = normal;
-                // A preserved zoom region is clipped to the new frame's domain; if
-                // it no longer intersects at all, fall back to the whole domain.
-                auto region = entry < spec.visibleRegions.size()
-                    && spec.visibleRegions[entry].has_value()
-                        ? *spec.visibleRegions[entry] : dataBounds;
-                for (int axis = 0; axis < metadata.dimension; ++axis) {
-                    const auto index = static_cast<std::size_t>(axis);
-                    auto lower = std::max(region.lower[index],
-                        dataBounds.lower[index]);
-                    auto upper = std::min(region.upper[index],
-                        dataBounds.upper[index]);
-                    if (!(lower < upper)) {
-                        lower = dataBounds.lower[index];
-                        upper = dataBounds.upper[index];
-                    }
-                    region.lower[index] = lower;
-                    region.upper[index] = upper;
-                }
-                request.visibleRegion = region;
-                request.outputSize = finestNativeOutputSize(
-                    metadata, request.visibleRegion, request.normalDirection);
-                request.composition = selectedLevel.composition;
-                request.maximumLevel = attemptMaximumLevel;
-                if (metadata.dimension == 3) {
-                    request.physicalPosition = positions[static_cast<std::size_t>(normal)];
-                }
-                auto display = executeSlice(result.dataset, request, rangeMode,
-                    spec.userRange, spec.logarithmic, spec.palette, cancellation);
-                display.mode = spec.displayMode;
-                display.contourCount = spec.contourCount;
-                if (isContourMode(spec.displayMode)) {
-                    appendContours(result.dataset, request, spec.contourCount,
-                        display.minimum, display.maximum, display.logarithmic,
-                        cancellation, display);
-                }
-                if (spec.displayMode == DisplayMode::VelocityVectors) {
-                    const auto u = std::min(spec.vectorUField, fieldCount - 1);
-                    const auto v = std::min(spec.vectorVField, fieldCount - 1);
-                    const auto w = std::min(spec.vectorWField, fieldCount - 1);
-                    auto [f1, f2] = (metadata.dimension == 3)
-                        ? (normal == 0 ? std::pair{v, w}
-                           : normal == 1 ? std::pair{u, w}
-                           : std::pair{u, v})
-                        : std::pair{u, v};
-                    display.vectorUField = f1;
-                    display.vectorVField = f2;
-                    appendVectorGlyphs(result.dataset, request,
-                        FieldId{f1}, FieldId{f2},
-                        spec.contourCount, cancellation, display);
-                }
-                result.displays.push_back(std::move(display));
-            }
-            // In 3-D, every views' "Visible" range must agree so the single color bar
-            // maps all three panels consistently. Compute the union of finite extrema
-            // across all three planes and re-render each display with the shared range.
-            if (result.displays.size() == 3 && rangeMode == RangeMode::Visible) {
-                double globalMin = std::numeric_limits<double>::infinity();
-                double globalMax = -std::numeric_limits<double>::infinity();
-                for (const auto& d : result.displays) {
-                    const auto range = finiteRange(d.slice.plane);
-                    if (range) {
-                        globalMin = std::min(globalMin, range->first);
-                        globalMax = std::max(globalMax, range->second);
-                    }
-                }
-                const auto logarithmic = spec.logarithmic;
-                if (!std::isfinite(globalMin) || !std::isfinite(globalMax)) {
-                    if (logarithmic) {
-                        globalMin = 1.0;
-                        globalMax = 10.0;
-                    } else {
-                        globalMin = 0.0;
-                        globalMax = 1.0;
-                    }
-                }
-                if (globalMin == globalMax) {
-                    if (logarithmic && globalMin > 0.0) {
-                        globalMin /= 1.0 + 1.0e-6;
-                        globalMax *= 1.0 + 1.0e-6;
-                    } else {
-                        const auto pad = std::max(std::abs(globalMin), 1.0) * 1.0e-6;
-                        globalMin -= pad;
-                        globalMax += pad;
-                    }
-                }
-                for (auto& d : result.displays) {
-                    d.minimum = globalMin;
-                    d.maximum = globalMax;
-                    d.image = renderScalarPlane(d.slice.plane,
-                        ScalarRenderSettings{
-                            .minimum = globalMin,
-                            .maximum = globalMax,
-                            .logarithmic = d.logarithmic,
-                            .palette = &spec.palette
-                        });
-                    // Contours were extracted per view before the shared range
-                    // was known; re-extract them so their levels match the
-                    // shared colorbar (see contours-stale-after-visible-range).
-                    recomputeContourPolylines(d);
-                }
-            }
-            break;
-        } catch (const CacheBudgetExceeded&) {
-            result.displays.clear();
-            result.dataset->clearUnpinnedCache();
-            if (selectedLevel.composition != CompositionPolicy::FinestAvailable) {
-                throw std::runtime_error(QObject::tr(
-                    "The selected slice level cannot fit in the %1 cache. "
-                    "Choose a lower level or increase AMREXPLORER_CACHE_SIZE_MB.")
-                        .arg(cacheBudgetDescription(cacheBudget))
-                        .toStdString());
-            }
-            if (attemptMaximumLevel == 0) {
-                throw std::runtime_error(QObject::tr(
-                    "The slice cannot fit in the %1 cache, even at level 0. "
-                    "Try a smaller plotfile or increase AMREXPLORER_CACHE_SIZE_MB.")
-                        .arg(cacheBudgetDescription(cacheBudget))
-                        .toStdString());
-            }
-            if (result.cacheFallbackFromLevel < 0) {
-                result.cacheFallbackFromLevel = attemptMaximumLevel;
-            }
-            result.cacheFallbackToLevel = --attemptMaximumLevel;
-        }
-    }
-    return result;
-}
 
 } // namespace
 
@@ -4335,7 +3595,8 @@ void MainWindow::requestInitialSlice(
         [path, generation, spec = std::move(spec), cancellation,
             preparedMetadata = std::move(preparedMetadata),
             dataRoot = std::move(dataRoot)]() mutable {
-        return executeFrameLoad(path, DatasetId{generation}, spec, cancellation,
+        return executeFrameLoad(path, DatasetId{generation}, spec,
+            initialCacheBudget(), cancellation,
             std::move(preparedMetadata), std::move(dataRoot));
     }));
 }
@@ -5433,7 +4694,8 @@ void MainWindow::startFrameLoad(int index, std::uint64_t generation)
         });
     watcher->setFuture(QtConcurrent::run(
         [path, datasetId, spec = std::move(spec), cancellation] {
-        return executeFrameLoad(path, datasetId, spec, cancellation);
+        return executeFrameLoad(path, datasetId, spec, initialCacheBudget(),
+            cancellation);
     }));
 }
 
@@ -5778,7 +5040,8 @@ void MainWindow::startPrefetch(int frameIndex)
         });
     watcher->setFuture(QtConcurrent::run(
         [path, datasetId, spec = std::move(spec), cancellation] {
-        return executeFrameLoad(path, datasetId, spec, cancellation);
+        return executeFrameLoad(path, datasetId, spec, initialCacheBudget(),
+            cancellation);
     }));
 }
